code/C1/requirements.txt
python-dotenv==1.0.1
Markdown==3.7
langchain==0.3.17
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.5
langchain-unstructured==0.1.6
langchain-deepseek-official==0.1.0
langchain-community==0.3.16
unstructured==0.16.17
modelscope==1.26.0
sentence-transformers==4.1.0
llama-index==0.12.15
llama-index-llms-deepseek==0.1.1
llama-index-embeddings-huggingface==0.5.4